// src/types/role-designation/roleDesignation.ts
// administration dropdown types
export type RoleDesigAdministrationDropDownTypes = {
  departmentID: string;
  name: string;
  email: string;
  phone: string;
  status: string;
};

// client dropdown types
export type RoleDesigClientDropDownTypes = {
  clientID: string;
  name: string;
};

// vendor dropdown types
export type RoleDesigVendorDropDownTypes = {
  vendorID: string;
  vendorSid: string;
  address: string;
  bin: string;
  tin: string;
  acNo: string;
  ownerName: string;
  typeOfBusiness: string;
  name: string;
  email: string;
  phone: string;
  status: string;
};